# src/testInt/resources/bean-validation/generated/model/Obj2.java
/*
 * This class is auto generated by https://github.com/hauner/openapi-generatr-spring.
 * DO NOT EDIT.
 */

package generated.model;

import com.fasterxml.jackson.annotation.JsonProperty;

import javax.validation.constraints.*;

public class Obj2 {

    @JsonProperty("prop4")
    @Size(max = 10)
    private String prop1;

    public String getProp4() {
        return prop4;
    }

    public void setProp4(String prop4) {
        this.prop4 = prop4;
    }

}