# src/testInt/resources/bean-validation/openapi.yaml
openapi: 3.0.2
info:
  title: test simple path parameters
  version: 1.0.0

paths:
  /endpoint/nullable:
    get:
      tags:
        - endpoint
      parameters:
        - in: query
          name: nullable-false
          schema:
            type: string
            nullable: false
        - in: query
          name: nullable-true
          schema:
            type: string
            nullable: true
      responses:
        '204':
          description: empty

  /endpoint/length:
    get:
      tags:
        - endpoint
      parameters:
        - in: query
          name: min-length
          schema:
            type: string
            minLength: 2
        - in: query
          name: max-length
          schema:
            type: string
            maxLength: 4
        - in: query
          name: min-max-length
          schema:
            type: string
            minLength: 2
            maxLength: 4
      responses:
        '204':
          description: empty

  /endpoint/minmax:
    get:
      tags:
        - endpoint
      parameters:
        - in: query
          name: min
          schema:
            type: integer
            format: int32
            minimum: 10
        - in: query
          name: min-ex
          schema:
            type: integer
            format: int32
            minimum: 10
            exclusiveMinimum: true
        - in: query
          name: max
          schema:
            type: integer
            format: int32
            maximum: 20
        - in: query
          name: max-ex
          schema:
            type: integer
            format: int32
            maximum: 20
            exclusiveMaximum: true
        - in: query
          name: min-max
          schema:
            type: integer
            format: int32
            minimum: 10
            maximum: 20
        - in: query
          name: min-max-ex
          schema:
            type: integer
            format: int32
            minimum: 10
            exclusiveMinimum: true
            maximum: 20
            exclusiveMaximum: true
      responses:
        '204':
          description: empty

  /endpoint/items:
    get:
      tags:
        - endpoint
      parameters:
        - in: query
          name: min
          schema:
            type: array
            minItems: 2
            items:
              type: string
        - in: query
          name: max
          schema:
            type: array
            maxItems: 4
            items:
              type: string
        - in: query
          name: min-max
          schema:
            type: array
            minItems: 2
            maxItems: 4
            items:
              type: string
      responses:
        '204':
          description: empty

  /endpoint/obj:
    post:
      tags:
        - endpoint
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Obj1'
      responses:
        '204':
          description: empty


components:
  schemas:
    Obj1:
      type: object
      properties:
        prop1:
          type: string
          nullable: false
        prop2:
          type: integer
          format: int32
          maximum: 3
        prop3:
          $ref: '#/components/schemas/Obj2'
    Obj2:
      type: object
      properties:
        prop4:
          type: string
          maxLength: 10